requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aerich==0.7.1
    # via -r requirements.in
aiofiles==23.1.0
    # via aiogram
aiogram==3.0.0b7
    # via -r requirements.in
aiohttp==3.8.4
    # via
    #   -r requirements.in
    #   aiogram
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.17.0
    # via tortoise-orm
alembic==1.11.1
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.27.0
    # via
    #   -r requirements.in
    #   tortoise-orm
attrs==22.2.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via bs4
bs4==0.0.1
    # via -r requirements.in
build==0.10.0
    # via pip-tools
cachetools==5.3.0
    # via -r requirements.in
certifi==2022.12.7
    # via aiogram
charset-normalizer==3.1.0
    # via aiohttp
click==8.1.3
    # via
    #   aerich
    #   pip-tools
dictdiffer==0.9.0
    # via aerich
exceptiongroup==1.1.2
    # via pytest
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via yarl
iniconfig==2.0.0
    # via pytest
iso8601==1.1.0
    # via tortoise-orm
jinja2==3.1.2
    # via -r requirements.in
load-dotenv==0.1.0
    # via -r requirements.in
lxml==4.9.2
    # via -r requirements.in
magic-filter==1.0.9
    # via aiogram
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.0
    # via
    #   build
    #   pytest
pillow==9.5.0
    # via -r requirements.in
pip-tools==6.13.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
pydantic==1.10.7
    # via
    #   -r requirements.in
    #   aerich
    #   aiogram
pypika-tortoise==0.1.6
    # via tortoise-orm
pyproject-hooks==1.0.0
    # via build
pytest==7.4.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   load-dotenv
pytz==2023.3
    # via tortoise-orm
six==1.16.0
    # via python-dateutil
soupsieve==2.4.1
    # via beautifulsoup4
sqlalchemy==2.0.15
    # via
    #   -r requirements.in
    #   alembic
tomli==2.0.1
    # via
    #   build
    #   pytest
tomlkit==0.11.8
    # via aerich
tortoise-orm[asyncpg]==0.19.3
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.5.0
    # via
    #   aiosqlite
    #   alembic
    #   pydantic
    #   sqlalchemy
wheel==0.40.0
    # via pip-tools
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
